pyproject.toml
[project]
name = "quick-convert"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "hyperpyyaml>=1.2.3",
    "omegaconf>=2.3.0",
    "soundfile>=0.13.1",
    "torch==2.6",
    "torchaudio==2.6",
    "transformers>=5.3.0",
    "tts>=0.22.0",
]
